feat: store edits per image
Previously, only the last edit was persisted, and applied
to any image that was subsequently opened.
By using a single UserDefault key per image, the parameters
are correctly stored per image.
Right now it uses a prefix followed by the name of the image.
A better solution would be to, e.g., compute an hash from the
original image content, and use that as the key to store the
filter parameter.