feat: apply the pixellate filter to the thumbnails
Note: this is a crude implementation, and applies the filter
only on a reload of the Gallery.
This means that after an Edit, the corresponding thumbnail
in the gallery is not updated until the app is relaunched.
One possible solution would be to filter the thumbnail
while editing the image, and pass the filtered one back
to the Gallery when exiting the Edit screen.